Relax matching on job name
Before this change job names were assumed to be in the format
`some_script.sh`, however this may not always be the case. For example,
when running MPI benchmarks the job name might be: `IMB-MPI1`. This
change makes the regex more flexible so that it can match a wider
range of job names.